feat: refactor origama and triton knobs
