+++ primus_turbo/triton/utils/triton_knobs_helper.py
@@ -1,0 +1,31 @@
+###############################################################################
+# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
+#
+# See LICENSE for license information.
+###############################################################################
+
+"""
+AMDGPU Triton compiler knob helpers (shared across GEMM / grouped-GEMM kernels).
+"""
+
+import os
+
+import triton
+
+_KNOBS_SET = False
+
+
+def set_triton_knobs_gfx950() -> None:
+    """Enable AMD compiler knobs for gfx950 (async_copy, block_pingpong, scalarize)."""
+    global _KNOBS_SET
+    if _KNOBS_SET:
+        return
+    _KNOBS_SET = True
+    if hasattr(triton, "knobs") and hasattr(triton.knobs, "amd"):
+        triton.knobs.amd.use_async_copy = True
+        triton.knobs.amd.scalarize_packed_fops = True
+        triton.knobs.amd.use_block_pingpong = True
+    else:
+        os.environ.setdefault("TRITON_HIP_USE_ASYNC_COPY", "1")
+        os.environ.setdefault("AMDGCN_SCALARIZE_PACKED_FOPS", "1")
+        os.environ.setdefault("TRITON_HIP_USE_BLOCK_PINGPONG", "1")
